refactor: using constants
relocate uplink token test to unit test
